Make the ACS endpoint configurable
The extension registers the SAML2 Assertion Consumer Service (ie the
endpoint that the IdP POSTs to after authenticating) as /acs, which is a
sensible default.

But existing applications using legacy extensions might already be using
a different ACS so this change make the actual route registered by CKAN
configurable.

There's is a check performed on startup to make sure the endpoint
configuration is a relative URL, ie starts with a slash ("/")