Support fullname field in SAMLResponse
As discussed in #11 in some cases it is useful to use the "full name"
claim rather than the concatenation of "first" + "last" name. Setting
aside that not all parts of the world use the first + last name format
for people's names, in some cases some customization is applied in the
IdP side to the user display name, eg to show the department the user
belongs to (eg "John Garcia (Operations)").

This change keeps all existing functionality relying on first + last
names but adds support for a new config option to provide the full name
instead. Startup checks and default fallbacks work in the same way as
before.