Support RelayState (redirecting to user-requested page)
This change adds support for redirecting the user to the original page a
user requested when they were redirected to the login page (ie same as
the `came_from` argument on the native login). This is useful
when the user is redirected to the login page after trying to access a
restricted one, specially on sites that require logged in users to
access any page at all.

In Saml2 this is implemented using the `RelayState` parameter, which is
sent by the client to the IdP when requesting the authentication, and
returned by the IdP to the ACS alongside the SamlResponse.