[#55] Clear CKAN cookies on SLO
As it is implemented now SLO correctly logs the user out from the IdP,
but leaves the local CKAN user logged in in the CKAN site. This is
because the auth and session cookies (`auth_tkt` and `ckan`) are not
cleared like in a normal logout.

This change modifies the response returned by the
`IAuthenticator.logout()` hook to include the necessary headers to clear
both cookies.