Fix broken login in CKAN>=2.9.6
CKAN 2.9.6 introduced a security change in the way Repoze auth cookies
are set up. Previously the user identifier was the user name, but that
has been changed to the user internal id plus a serial autoincrement
(currently static).
As ckanext-saml2auth calls the `set_repoze_user` directly we need to
update what gets passed to be stored in the cookie, otherwise the user
will never be logged in.
Previous versions should keep using the user name to avoid the same
problem.

Note that all this will no longer be relevant on CKAN 2.10 as it uses a
completely different login system.